Altered test folder structure
